fix #1881: Open generatorUrl in a new browser tab.
There does not appear to be a good way to reliably determine in
onUrlRequest whether the URL being navigated to is an Internal or
External URL, especially if Alertmanager is hosted on the same domain as
other services (for example, Prometheus). To workaround this problem for
the specific case of the generatorUrlButton (labeled "Source"), we force
it to have target="_blank", which prevents Elm from intercepting the URL
and trying to do client-side navigation.

Signed-off-by: Daniel Pryden <daniel@pryden.net>